treewide: help: Add better case-insensitive help command implementation
With the previous implementations, argument was being titled and compared with module names in the help dictionary. This made modules with non-standart letter case(such as PMPermit) fail when queryed for help page. This new implementation ensures that the correct help page is given, even if the case doesn't match, and it ensures the module name is properly written in the Help page itself(with the casing preserved, such as 'PMPermit' instead of 'pmpermit').

Signed-off-by: Gegham Zakaryan <zakaryan.2004@outlook.com>