filters: Mark chat as unread, save link to unread messages, fixes
* Sending a message also sends read acknowledgement. To circumvent
this, when a filter is triggered and BOTLOG is enabled, Paperplane
will create a link to the next unread message in the chat and
send it to BOTLOG chat, so the user can jump back to it easily.

* Additionally, that chat will be marked as unread, so the user
can realize that a Paperplane filter was triggered there.

* Finally, the .filter command has been fixed to take other types
of double quotes into account, since Telegram likes to replace the
conventional ones with other characters.

Signed-off-by: Gegham Zakaryan <zakaryan.2004@outlook.com>